+++ src/main.ts
@@ -1,6 +1,7 @@
 import { ValidationPipe } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import * as bodyParser from 'body-parser';
+import helmet from 'helmet';
 import { AppModule } from './app.module';
 import { AllExeptionsFilter } from './filters/all-exeptions.filter';
 
@@ -9,23 +10,20 @@ import { AllExeptionsFilter } from './filters/all-exeptions.filter';
 async function bootstrap() {
   const isProd = process.env.NODE_ENV === 'production';
 
-  //Use without nginx
-  // const app = isProd
-  //   ? await NestFactory.create(AppModule, {
-  //       httpsOptions: {
-  //         key: fs.readFileSync(process.env.SSL_KEY_PATH || './ssl/key.key'),
-  //         cert: fs.readFileSync(process.env.SSL_CERT_PATH || './ssl/cert.pem'),
-  //         ca: fs.readFileSync(process.env.SSL_CA_PATH || './ssl/ca_bundle.pem'),
-  //       },
-  //     })
-  //   : await NestFactory.create(AppModule);
-
   const app = await NestFactory.create(AppModule);
+  //security headers
+  app.use(helmet({
+    contentSecurityPolicy: false
+  }));
 
   app.use(bodyParser.json({ limit: '5mb' }));
   app.use(bodyParser.urlencoded({ limit: '5mb', extended: true }));
+
   //root uri
   app.setGlobalPrefix('api');
+
+  // global guard for jwt auth
+  //app.useGlobalGuards(new JwtAuthGuard());
 
   // enabled cores from any dispositive
   app.enableCors({